+++ src/Core/Dfe.Complete.Application/Contacts/Models/KeyContactsDto.cs
@@ -1,0 +1,20 @@
+﻿using Dfe.Complete.Domain.ValueObjects;
+
+namespace Dfe.Complete.Application.Contacts.Models
+{
+    public class KeyContactsDto
+    {
+        public KeyContactId? Id { get; set; }
+        public ContactId? HeadteacherId { get; set; }
+
+        public DateTime CreatedAt { get; set; }
+
+        public DateTime UpdatedAt { get; set; }
+
+        public ContactId? ChairOfGovernorsId { get; set; }
+
+        public ContactId? IncomingTrustCeoId { get; set; }
+
+        public ContactId? OutgoingTrustCeoId { get; set; }
+    }
+}