+++ src/Core/Dfe.Complete.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -41,6 +41,7 @@ namespace Dfe.Complete.Infrastructure
             services.AddScoped<INoteReadRepository, NoteReadRepository>();
             services.AddScoped<IReadUserRepository,ReadUserRepository>();
             services.AddScoped<ILocalAuthoritiesQueryService, LocalAuthoritiesQueryService>();
+            services.AddScoped<ITaskDataReadRepository, TaskDataReadRepository>();
 
             // Authentication
             //services.AddCustomAuthorization(config);