+++ src/Frontend/Dfe.Complete/Models/TaskListStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace Dfe.Complete.Models
+{
+    public enum TaskListStatus
+    {
+        NotApplicable,
+        NotStarted,
+        InProgress,
+        Completed
+    }
+}