+++ CHANGELOG.md
@@ -17,8 +17,10 @@ Security in case of vulnerabilities.
 ## Unreleased
 
 ### Added
+
 - Added Project Significant History Dates `/projects/{projectId}/date-history`
 - Added `/projects/{projectId}/external-contacts` page
+- Added `Conversion` and `Transfer` task list with their statuses.
 
 ### Changed